+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED]
+### Added
+- Add silent option for `tabs.navToTab` method
+
 ## [0.6.1] - 2018-08-27
 ### Added
 - Tabs API can now be observed